smart-api-optimizer/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0
aiosqlite==0.19.0

# Database Migrations
alembic==1.12.1

# Redis & Caching
redis==5.0.0
redis[hiredis]==5.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
passlib==1.7.4

# Background Jobs
celery==5.3.4
click==8.1.7

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Environment & Configuration
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6
